requirements.txt
fastapi
uvicorn
streamlit
groq
langchain
langchain-community
langchain-huggingface
langchain-chroma
chromadb
sentence-transformers
rank_bm25
pdfplumber
torch
python-dotenv
pydantic
numpy
scikit-learn
unstructured
markdown
